refactor(examples): fold A2A example into customer-service
Move the A2A tool call capture example from the feature-named
examples/tool-call-capture/a2a/ into examples/customer-service/a2a_server/
so all three transport variants (Python connector, traced, A2A) share
one domain, one set of scenarios, and one tool suite.

- Extract shared SQLite tools into tools.py (6 tools, DB setup,
  agent instructions) imported by all three agent variants
- Gate tool call emission on A2A-Extensions request header per
  least-privilege (server only emits when client requests it)
- Drop artifact name="response" from docs snippet (not contractual)
- Rephrase DataPart warning to neutral spec-anchored tone
- Add Security section to tool-call-capture.mdx covering threat
  model and three deployment options (dedicated, auth, env flag)
- Delete examples/tool-call-capture/ and update all references
