+++ examples/customer-service/config_a2a.yaml
@@ -1,64 +1,47 @@
-# Simulate + Evaluate Configuration File
-# Configuration for 'arksim simulate-evaluate' command
-# Runs simulation followed by evaluation in a single pipeline.
-#
-# Weather agent that reports tool calls via the arksim A2A tool capture
-# extension (https://arksim.arklex.ai/a2a/tool-call-capture/v1).
-# Tools: get_weather
-# arksim reads tool calls from Task.artifacts[*].metadata and evaluates
-# tool usage against the assertions in scenarios.json.
+# A2A Connector Configuration
+# Connects to the customer-service agent via the A2A protocol.
+# The agent surfaces tool calls via the arksim A2A tool capture extension
+# (https://arksim.arklex.ai/a2a/tool-call-capture/v1). arksim reads tool
+# calls from Task.artifacts[*].metadata and evaluates tool usage against
+# the assertions in scenarios.json.
 #
 # Usage:
-#   cd examples/tool-call-capture/a2a
-#   python -m agent_server.server &
-#   arksim simulate-evaluate config.yaml
+#   cd examples/customer-service
+#   python -m a2a_server.server &
+#   arksim simulate-evaluate config_a2a.yaml
 
 # ── AGENT CONFIGURATION ────────────────────────────────────────────────────
 
 agent_config:
   agent_type: a2a
-  agent_name: weather-tool-capture
+  agent_name: customer-service-a2a
   api_config:
     endpoint: http://localhost:9998
 
 # ── SIMULATION SETTINGS ──────────────────────────────────────────────────────
 
-# Path to the scenarios file
 scenario_file_path: ./scenarios.json
-
-# Number of conversations per scenario to generate
 num_conversations_per_scenario: 1
-
-# Maximum turns per conversation
-max_turns: 3
-
-# Output file path for simulation results
-output_file_path: ./results/simulation/simulation.json
-
-# Jinja template for the simulated user's system prompt
+max_turns: 5
+output_file_path: ./results/simulation/simulation_a2a.json
 simulated_user_prompt_template: null
 
 # ── EVALUATION SETTINGS ──────────────────────────────────────────────────────
 
-# Output directory for evaluation results
 output_dir: ./results/evaluation
 
-# Built-in metrics to run; if empty, all built-in metrics run
+custom_metrics_file_paths:
+  - ./custom_metrics.py
+
 metrics_to_run:
   - helpfulness
   - agent_behavior_failure
   - tool_call_behavior_failure
 
-# Generate HTML report
 generate_html_report: true
 
 # ── SHARED SETTINGS ──────────────────────────────────────────────────────────
 
-# LLM model
 model: gpt-4.1-mini
-
-# LLM provider
 provider: openai
-
-# Workers for parallel processing
 num_workers: 1
