--- examples/tool-call-capture/a2a/requirements.txt
@@ -1,4 +1,0 @@
-arksim>=0.3.4
-openai-agents>=0.0.17
-a2a-sdk>=0.3.20
-uvicorn>=0.38.0
